Rename value_balance_to_point to transaction_fee_to_point
Also added a note explaining that this function should stay specific to transaction fee because of the introduction of asset type. Transaction fees should always be denominated in the default asset type, so if this function is needed for other places, it's probably best to create a similar one that takes in an AssetType as a parameter such that we don't end up introducing a bug inadvertently.